redir_in works; logic apparently doesn't

+++ sources/execution/heredoc.c
@@ -5,6 +5,8 @@ static void count_heredocs(t_ast *s, t_exe *b)
     t_ast   *subsh_ast;
     t_list  *ins_cpy;
 
+    if (!s || !b)
+        return ;
     subsh_ast = NULL;
     if (s->tag == COMMAND && s->command->ins)
     {
@@ -20,7 +22,7 @@ static void count_heredocs(t_ast *s, t_exe *b)
     {
         subsh_ast = parse(s->subshell_cmd, b->env);
         if (!subsh_ast)
-            ft_putstr_fd("count heredocs: parse: malloc fail\n", 2);
+            ft_putstr_fd("count heredocs: recursive parsing fail\n", 2);
         else if (subsh_ast)
             count_heredocs(subsh_ast, b);
         free_ast(subsh_ast);
@@ -50,10 +52,10 @@ static void fetch_hd_delimiters(t_ast *s, t_exe *b)
     if (s->tag == SUBSHELL || s->tag == RECCALL)
     {
         subsh_ast = parse(s->subshell_cmd, b->env);
-        if (subsh_ast->left)
-            fetch_hd_delimiters(subsh_ast->left, b);
-        if (subsh_ast->right)
-            fetch_hd_delimiters(subsh_ast->right, b);
+        if (!subsh_ast)
+            ft_putstr_fd("fetch_hd_delimiters: recursive parsing fail\n", 2);
+        else if (subsh_ast)
+            fetch_hd_delimiters(subsh_ast, b);
         free_ast(subsh_ast);
     }
     if (s->left)
@@ -74,12 +76,31 @@ static void open_files_for_heredocs(t_exe *exe_bus)
         num = ft_itoa(i);
         path = ft_strjoin("/tmp/heredoc", num);
         free(num);
-        exe_bus->hd_fds[i] = open(path, O_RDWR | O_CREAT | O_APPEND, 0644);
+        exe_bus->hd_fds[i] = open(path, O_WRONLY| O_CREAT | O_TRUNC, 0644);
         if (exe_bus->hd_fds[i] < 0)
             perror("heredoc: ");
         free(path);
     }
 }
+
+// static void reopen_hd_files_for_reading(t_exe *exe_bus)
+// {
+//     int     i;
+//     char    *num;
+//     char    *path;
+
+//     i = -1;
+//     while (++i < exe_bus->hd_count)
+//     {
+//         num = ft_itoa(i);
+//         path = ft_strjoin("/tmp/heredoc", num);
+//         free(num);
+//         exe_bus->hd_fds[i] = open(path, O_RDONLY, 0644);
+//         if (exe_bus->hd_fds[i] < 0)
+//             perror("heredoc: ");
+//         free(path);
+//     }
+// }
 
 static void prompt_for_one_heredoc(t_exe *exe_bus, char *limiter, int i)
 {
@@ -129,6 +150,8 @@ static void prompt_for_all_heredocs(t_exe *exe_bus)
             continue ;
         }
         prompt_for_one_heredoc(exe_bus, limiter, i);
+        if (close(exe_bus->hd_fds[i]) < 0)
+            ft_putstr_fd("heredoc after writing: close fail\n", STDERR_FILENO);
         free(limiter);
     }
 }
@@ -148,8 +171,9 @@ void    exec_heredocs(t_exe *exe_bus)
     open_files_for_heredocs(exe_bus);
     // run prompts
     prompt_for_all_heredocs(exe_bus);
+    // reopen_hd_files_for_reading(exe_bus);
     // free
-    free_heredocs(exe_bus); // just for leak control now, will actually be needed in redir
+    // free_heredocs(exe_bus); // just for leak control now, will actually be needed in redir
 }
 
 void    free_heredocs(t_exe *exe_bus)
@@ -159,12 +183,14 @@ void    free_heredocs(t_exe *exe_bus)
     char    *path;
 
     i = -1;
+    if (exe_bus->hd_count == 0)
+        return ;
     while (++i < exe_bus->hd_count)
     {
         num = ft_itoa(i);
         path = ft_strjoin("/tmp/heredoc", num);
         free(num);
-        close(exe_bus->hd_fds[i]);
+        // close(exe_bus->hd_fds[i]);
         unlink(path);
         free(path);
         free(exe_bus->hd_delimiters[i]);
